services and users custom hooks
