+++ src/components/settings/services/CreateServiceModal/CreateServiceModal.tsx
@@ -1,7 +1,5 @@
 "use client"
 
-import { useEffect, useState } from "react"
-import { z } from "zod"
 import {
   CloseButton,
   DialogActions,
@@ -10,6 +8,7 @@ import {
   DialogTitle,
   Modal,
 } from "@/components/common/dialog"
+import { useCreateServiceModal } from "@/hooks/settings/services/useCreateServiceModal"
 import type { LocationWithRelations } from "@/lib/prisma/location/types"
 import type { ServiceWithRelations } from "@/lib/prisma/service/types"
 import type { ServiceCategoryWithRelations } from "@/lib/prisma/service_category/types"
@@ -34,105 +33,16 @@ export const CreateServiceModal = ({
   doesServiceCodeExist,
   revalidateTable,
 }: CreateServiceModalProps) => {
-  const [isSaving, setIsSaving] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [formData, setFormData] = useState<Partial<ServiceWithRelations> | null>(null)
-  const [isFormValidState, setIsFormValidState] = useState<boolean>(false)
-  const [isFormValidating, setIsFormValidating] = useState<boolean>(false)
-
-  // initialize form data when the modal opens
-  useEffect(() => {
-    if (open) {
-      setFormData({
-        name: "",
-        code: "",
-        description: "",
-        publicName: "",
-        ticketPrefix: "",
-        legacyServiceId: null,
-        backOffice: false,
-        deletedAt: null,
-        locations: [],
-        categories: [],
-      })
-    } else {
-      setFormData(null)
-      setIsFormValidState(false)
-      setIsFormValidating(false)
-    }
-  }, [open])
-
-  const NewServiceWithRelationsSchema = z.object({
-    name: z.string().min(1, "Name is required"),
-    code: z
-      .string()
-      .min(1, "Code is required")
-      .refine(
-        async (code) => {
-          return !(await doesServiceCodeExist(code))
-        },
-        { message: "Code already exists" }
-      ),
-    description: z.string(),
-    publicName: z.string().min(1, "Public name is required"),
-    ticketPrefix: z.string().min(1, "Ticket prefix is required"),
-    legacyServiceId: z.number().nullable(),
-    backOffice: z.boolean(),
-    locations: z.array(z.any()),
-  })
-
-  // Validate formData asynchronously and update local state instead of calling async validators during render
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <>
-  useEffect(() => {
-    if (!formData) {
-      setIsFormValidState(false)
-      setIsFormValidating(false)
-      return
-    }
-
-    let active = true
-    setIsFormValidating(true)
-
-    NewServiceWithRelationsSchema.parseAsync(formData)
-      .then(() => {
-        if (active) setIsFormValidState(true)
-      })
-      .catch(() => {
-        if (active) setIsFormValidState(false)
-      })
-      .finally(() => {
-        if (active) setIsFormValidating(false)
-      })
-
-    return () => {
-      active = false
-    }
-  }, [formData, doesServiceCodeExist])
+  const { isSaving, error, formData, setFormData, isReadonly, isSaveDisabled, handleSave } =
+    useCreateServiceModal({
+      open,
+      onClose,
+      insertService,
+      doesServiceCodeExist,
+      revalidateTable,
+    })
 
   if (!formData) return null
-
-  const isArchived = formData.deletedAt !== null
-  const isReadonly = isArchived
-
-  const handleSave = async () => {
-    if (formData && !isReadonly) {
-      try {
-        setIsSaving(true)
-        await insertService(formData)
-        await revalidateTable()
-        onClose()
-        setIsSaving(false)
-        window.location.href = "/protected/settings/services"
-      } catch (e: unknown) {
-        if (e instanceof Error) {
-          setError(e.message)
-        } else {
-          setError("An unknown error occurred")
-        }
-        setIsSaving(false)
-      }
-    }
-  }
 
   return (
     <Modal open={open} onClose={onClose} size="lg">
@@ -163,12 +73,7 @@ export const CreateServiceModal = ({
         <button type="button" className="tertiary" onClick={onClose}>
           Cancel
         </button>
-        <button
-          type="button"
-          className="primary"
-          onClick={handleSave}
-          disabled={isSaving || isFormValidating || !isFormValidState}
-        >
+        <button type="button" className="primary" onClick={handleSave} disabled={isSaveDisabled}>
           {isSaving ? "Saving..." : "Save Changes"}
         </button>
       </DialogActions>
