+++ src/components/settings/services/ServiceTable/ServiceTable.tsx
@@ -1,15 +1,12 @@
 "use client"
 
-import { useEffect, useMemo, useState } from "react"
 import { DataTable } from "@/components/common/datatable"
 import { Switch } from "@/components/common/switch"
-import { useAuth, useDialog } from "@/hooks"
+import { useServiceTable } from "@/hooks/settings/services/useServiceTable"
 import type { LocationWithRelations } from "@/lib/prisma/location/types"
 import type { ServiceWithRelations } from "@/lib/prisma/service/types"
 import type { ServiceCategoryWithRelations } from "@/lib/prisma/service_category/types"
 import type { StaffUserWithRelations } from "@/lib/prisma/staff_user/types"
-import { resolvePolicy } from "@/utils/policies/resolvePolicy"
-import type { UserContext } from "@/utils/policies/types"
 import { ConfirmArchiveServiceModal } from "../ConfirmArchiveServiceModal"
 import { CreateServiceModal } from "../CreateServiceModal"
 import { EditServiceModal } from "../EditServiceModal"
@@ -39,66 +36,24 @@ export const ServiceTable = ({
   doesServiceCodeExist,
   revalidateTable,
 }: ServiceTableProps) => {
-  const { role, idir_user_guid } = useAuth()
   const {
-    open: editServiceModalOpen,
-    openDialog: openEditServiceModal,
-    closeDialog: closeEditServiceModal,
-  } = useDialog()
-  const {
-    open: createServiceModalOpen,
-    openDialog: openCreateServiceModal,
-    closeDialog: closeCreateServiceModal,
-  } = useDialog()
-  const {
-    open: confirmArchiveServiceModalOpen,
-    openDialog: openConfirmArchiveServiceModal,
-    closeDialog: closeConfirmArchiveServiceModal,
-  } = useDialog()
-
-  const userContext = useMemo<UserContext>(
-    () => ({
-      staff_user_id: idir_user_guid ?? null,
-      role,
-      location_code: currentUser?.locationCode ?? null,
-    }),
-    [idir_user_guid, role, currentUser?.locationCode]
-  )
-
-  const actions = resolvePolicy("service", userContext)
-
-  const [showArchived, setShowArchived] = useState<boolean>(false)
-  const [selectedService, setSelectedService] = useState<ServiceWithRelations | null>(null)
-  const [canEditSelectedService, setCanEditSelectedService] = useState<boolean>(false)
-  const [canArchiveSelectedService, setCanArchiveSelectedService] = useState<boolean>(false)
-
-  // Determine if the current user can edit/archive the selected service whenever either changes
-  useEffect(() => {
-    if (selectedService) {
-      const actions = resolvePolicy("service", userContext, selectedService)
-      setCanEditSelectedService(actions.includes("edit"))
-      setCanArchiveSelectedService(actions.includes("archive"))
-    } else {
-      setCanEditSelectedService(false)
-      setCanArchiveSelectedService(false)
-    }
-  }, [selectedService, userContext])
-
-  const handleRowClick = (service: ServiceWithRelations) => {
-    setSelectedService(service)
-    openEditServiceModal()
-  }
-
-  const servicesToShow = services.filter((service) => {
-    // Filter by archived status
-    if (!showArchived && service.deletedAt !== null) return false
-
-    // Filter by view permission
-    const actions = resolvePolicy("service", userContext, service)
-    return actions.includes("view")
-  })
-
-  const canCreate = actions.includes("create")
+    showArchived,
+    setShowArchived,
+    selectedService,
+    canCreate,
+    canEditSelectedService,
+    canArchiveSelectedService,
+    servicesToShow,
+    handleRowClick,
+    editServiceModalOpen,
+    closeEditServiceModal,
+    createServiceModalOpen,
+    openCreateServiceModal,
+    closeCreateServiceModal,
+    confirmArchiveServiceModalOpen,
+    openConfirmArchiveServiceModal,
+    closeConfirmArchiveServiceModal,
+  } = useServiceTable({ currentUser, services, revalidateTable })
 
   return (
     <>
