+++ src/hooks/settings/services/useConfirmArchiveServiceModal/index.ts
@@ -1,0 +1,1 @@
+export { useConfirmArchiveServiceModal } from "./useConfirmArchiveServiceModal"
