+++ src/hooks/settings/services/useEditServiceModal/index.ts
@@ -1,0 +1,1 @@
+export { useEditServiceModal } from "./useEditServiceModal"
