+++ src/hooks/settings/services/useServiceTable/index.ts
@@ -1,0 +1,1 @@
+export { useServiceTable } from "./useServiceTable"
