+++ src/hooks/settings/users/useConfirmArchiveUserModal/index.ts
@@ -1,0 +1,1 @@
+export { useConfirmArchiveUserModal } from "./useConfirmArchiveUserModal"
